requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
pytz==2023.3
pydantic==2.3.0
psycopg2-binary==2.9.9